add core module
